Put the am rendezvous in ucpx.h. rather than ucp.h
